[script.domoticz.scenes] 0.0.8
